fix(product-add): url 수정 및 entity 수정
- url /api 빠진 것 수정
- Product.stock long에서 int로 변경

+++ src/main/java/com/example/onionstore/domain/product/dto/ProductCreateRequest.java
@@ -17,6 +17,6 @@ public record ProductCreateRequest(
         long price,
 
         @Min(1)
-        long stock
+        int stock
 ) {
 }
